refactor(raster-tile-layer): Colocate props type with class; guard level lookup
- Merge types.ts into raster-tile-layer.ts to match deck.gl convention of
  "props type then layer class" in the same file.
- Guard _renderSubLayers against out-of-range tile zoom levels instead of
  asserting descriptor.levels[z] is defined.
- Drop redundant `| undefined` from accessor return types — the prop types
  already express optionality.
- Remove empty initializeState override; CompositeLayer initializes state
  to {} by default.

Addresses post-commit review of d08d237.
